Update PWSQC_1.R
Initially the code made a projection of the lon and lat locations (in rijksdriehoekscoordinaten) in order to calculate distances between stations in meter. The projection that was used is appropriate for the Netherlands, but not for other areas in the world. 
These lines in the code have been replaced with a function from the geosphere package, which calculates distance between stations directly from their lon and lat locations. The function (distHaversine) assumes a spherical earth, ignoring ellipsoidal effects. The new code is applicable on PWS networks anywhere in the world. 

Courtesy of Dr. Maarten Reyniers and Eva Beele.