Add files via upload
For very large PWS networks, the code could be made more efficient, by changing some functions in the neighbourselection in script PWSQC_1.R and limiting the number of neighbours to a fixed maximum. More information can be found as notes in the script. Courtesy of Dr. Aart Overeem and Niek van Andel.